requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.10
configparser==5.2.0
idna==3.3
playsound==1.3.0
pydub==0.25.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
urllib3==1.26.9
